requirements.txt
asgiref==3.4.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.5.7
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
commandlines==0.4.1
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
cycler==0.10.0
daphne==3.0.2
Django==3.2.8
django-cleanup==5.2.0
django-countries==7.2.1
django-extensions==3.1.3
django-js-asset==1.2.2
django-leaflet==0.28.1
django-map-widgets==0.3.2
django-mptt==0.13.4
django-restframework==0.0.1
djangorestframework==3.12.4
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
filetype==1.0.8
Flask==2.0.2
GDAL==3.3.2
gunicorn==20.1.0
h11==0.12.0
hsh==1.1.0
hyperlink==21.0.0
idna==3.3
imageio==2.9.0
incremental==21.3.0
inflection==0.5.1
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.0.2
jsonschema==4.1.2
kiwisolver==1.3.2
llvmlite==0.37.0
MarkupSafe==2.0.1
matplotlib==3.4.3
msgpack-python==0.5.6
networkx==2.6.3
numba==0.54.1
numpy==1.20.3
packaging==21.0
Pillow==8.4.0
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==2.3.0
PyMatting==1.1.3
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.1.1
PyYAML==6.0
redis==3.5.3
rembg==1.0.27
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
scikit-image==0.18.3
scipy==1.7.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
swagger-spec-validator==2.7.3
tifffile==2021.10.12
toml==0.10.2
torch==1.10.0
torchvision==0.11.1
tqdm==4.62.3
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
uvicorn==0.15.0
waitress==2.0.0
Werkzeug==2.0.2
zope.interface==5.4.0
